Added OpenBSD support.
Bump _XOPEN_SOURCE to 800 for mkostemp (part of POSIX.1-2024).

Define _BSD_SOURCE in drm.c for major.

Guard use of posix_fallocate on _POSIX_ADVISORY_INFO. If it is -1
or undefined, the Advisory Information option is not supported for
compilation.

Always link with -l c to prevent undefined symbol errors with lld.

Co-authored-by: Michael Forney <mforney@mforney.org>